dts: stm32: add adc2 to stm32mp131
Add STM32 ADC2 support to STM32MP131 SoCs family.

VBAT internal channel is not defined in SoC DT. It has to be defined
in board DT, when needed, instead. This avoids unwanted
current consumption on battery, when ADC conversions are performed
on any other channels.

Set ADC2 as a child node of ETZPC, as the ADC2 accesses must be controlled
by the firewall.

Depends-on: I70556a98c3de24a7a509912b1dae0600291493fb

Signed-off-by: Olivier Moysan <olivier.moysan@foss.st.com>
Change-Id: Ib5fc4b27bf70b80c9fd2792e206fb179ef7e631c
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/247587
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Fabrice GASNIER <fabrice.gasnier@foss.st.com>
Reviewed-by: Gatien CHEVALLIER <gatien.chevallier@st.com>