drivers: stm32_huk: read a HUK from OTP
For STM32MP15, provide a HUK from OTP. If OTP is not provisioned,
return an error.
We define CFG_OTP_HW_TESTKEY=y in conf.mk to be able to boot without
error on a not provisioned board. In this case the HUK is statically
defined key, and the same for all soc.

A production product must use CFG_OTP_HW_TESTKEY=n.

Signed-off-by: Nicolas Toromanoff <nicolas.toromanoff@foss.st.com>
Change-Id: I58f2427191e70b0ce00520d2fdc47c26eac39e17
Reviewed-on: https://gerrit.st.com/c/mpu/oe/optee/optee_os/+/265211
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Gatien CHEVALLIER <gatien.chevallier@st.com>
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Nicolas TOROMANOFF <nicolas.toromanoff@st.com>
Tested-by: Nicolas TOROMANOFF <nicolas.toromanoff@st.com>